Refactor Works and ProjectsPage components, move categories filter to ProjectsPage, remove categories filter from Works
