+++ src/container/Works/Works.jsx
@@ -14,58 +14,20 @@ import './Works.scss';
 import { MoreLink } from '../../components';
 
 const Works = () => {
-  const { works, projects } = useSanityContext();
-  const [activeFilter, setActiveFilter] = useState('All');
-  const [animateCard, setAnimateCard] = useState({ y: 0, opacity: 1 });
-  const [filterWorks, setFilterWork] = useState([]);
-
-  useEffect(() => {
-    if (projects && projects.length > 0) {
-      setFilterWork(projects);
-    }
-  }, [projects]);
-
-  const handleWorkFilter = (item) => {
-    setActiveFilter(item);
-    setAnimateCard({ y: 100, opacity: 0 });
-    setTimeout(() => {
-      setAnimateCard({ y: 0, opacity: 1 });
-      if (item === 'All') {
-        setFilterWork(works);
-      } else {
-        const filter = works.filter((work) => work.tags.includes(item));
-        setFilterWork(filter);
-      }
-    }, 500);
-  };
+  const { projects } = useSanityContext();
 
   return (
     <>
       <h2 className="head-text">
         Things I've <span>Made</span>
       </h2>
-      <div className="app__work-filter">
-        {['UI/UX', 'Web App', 'Mobile App', 'React JS', 'All'].map(
-          (item, index) => (
-            <div
-              key={index}
-              onClick={() => handleWorkFilter(item)}
-              className={`app__work-filter-item app__flex p-text ${
-                activeFilter === item ? 'item-active' : ''
-              }`}
-            >
-              {item}
-            </div>
-          )
-        )}
-      </div>
+
       <motion.div
-        animate={animateCard}
         transition={{ duration: 0.5, delayChildren: 0.5 }}
         className="app__work-portfolio"
       >
-        {filterWorks.slice(0, 3).map((work, index) => (
-          <Work key={index} project={work} hoverLinks />
+        {projects.slice(0, 3).map((proj) => (
+          <Work key={proj._id} project={proj} hoverLinks />
         ))}
       </motion.div>
       <MoreLink text="View More" path="/projects" />
