+++ src/pages/ProjectsPage.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 
 import { Link } from 'react-router-dom';
 import { useSanityContext } from '../context/SanityContext';
@@ -9,7 +9,29 @@ import { images } from '../constants/';
 import './ProjectsPage.scss';
 
 const ProjectsPage = () => {
-  const { projects, works } = useSanityContext();
+  const { projects, projectCategories } = useSanityContext();
+  const [activeFilter, setActiveFilter] = useState('All');
+  const [filterProjects, setFilterProjects] = useState([]);
+  const [animateCard, setAnimateCard] = useState({ y: 0, opacity: 1 });
+
+  useEffect(() => {
+    if (projects && projects.length > 0) {
+      setFilterProjects(projects);
+    }
+  }, [projects]);
+
+  const handleWorkFilter = (item) => {
+    setActiveFilter(item.name);
+    setAnimateCard({ y: 100, opacity: 0 });
+    setTimeout(() => {
+      setAnimateCard({ y: 0, opacity: 1 });
+
+      const filter = projects.filter((proj) =>
+        proj.categories.some((category) => category.name === item.name)
+      );
+      setFilterProjects(filter);
+    }, 500);
+  };
   return (
     <div className="app__container col">
       <ImageHeading
@@ -19,8 +41,21 @@ const ProjectsPage = () => {
         height={350}
       />
       <section>
+        <div className="projects__project-filter">
+          {projectCategories.map((item) => (
+            <div
+              key={item._id}
+              onClick={() => handleWorkFilter(item)}
+              className={`projects__project-filter-item app__flex p-text ${
+                activeFilter === item.name ? 'item-active' : ''
+              }`}
+            >
+              {item.name}
+            </div>
+          ))}
+        </div>
         <div className="projects__container app__work-portfolio">
-          {projects.map((project, index) => (
+          {filterProjects.map((project, index) => (
             <Link to={`/projects/${project.slug.current}`} key={project._id}>
               <Work project={project} hoverLinks={false} />
             </Link>
