+++ src/pages/ProjectsPage.scss
@@ -1,6 +1,45 @@
-
-
-
 .projects__container a {
     text-decoration: none;
+}
+
+.projects__project-filter {
+    display: flex;
+    flex-direction: row-reverse;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+
+    margin: 4rem 0 2rem;
+.projects__project-filter-item {
+    padding: 0.5rem 1rem;
+    border-radius: 0.5rem;
+    background-color: var(--text-white);
+    color: var(--text);
+    font-weight: 600;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    margin: 0.5rem;
+
+    &:hover {
+        background-color: var(--secondary);
+        color: var(--text-white);
+    }
+
+    @media screen and (min-width: 2000px) {
+        padding: 1rem 2rem;
+        border-radius: 0.85rem;
+        
+    }
+
+}
+.item-active {
+    background-color: var(--secondary);
+    color: var(--text-white);
+}
+
+@media screen and (max-width: 900px) {
+    justify-content: center;
+    margin: 2rem 0 1rem;
+    
+}
 }
